Move vtcombo initTabletMap func to vtcombo package
The idea is so another commands can reuse the initTabletMap.

In my case, I am building a new command and importing libs from vitess. In my cmd I am
faking a topology(using memory) and starting vttablets for it. Similar on what vtcombo
does.
